Restore bookkeeper config changes by moving PrometheusMetricsProvider under a different package
