+++ conf/bookkeeper.conf
@@ -595,7 +595,7 @@ compactionRateByBytes=1000000
 # enableStatistics=true
 
 # Stats Provider Class (if statistics are enabled)
-statsProviderClass=org.apache.pulsar.metrics.prometheus.bookkeeper.PrometheusMetricsProvider
+statsProviderClass=org.apache.pulsar.bookkeeper.stats.prometheus.PrometheusMetricsProvider
 
 # Default port for Prometheus metrics exporter
 prometheusStatsHttpPort=8000
