+++ jetty-upgrade/bookkeeper-prometheus-metrics-provider/src/main/java/org/apache/pulsar/metrics/prometheus/bookkeeper/PrometheusStatsLogger.java
@@ -26,6 +26,7 @@ import org.apache.bookkeeper.stats.Counter;
 import org.apache.bookkeeper.stats.Gauge;
 import org.apache.bookkeeper.stats.OpStatsLogger;
 import org.apache.bookkeeper.stats.StatsLogger;
+import org.apache.pulsar.bookkeeper.stats.prometheus.PrometheusMetricsProvider;
 
 /**
  * A {@code Prometheus} based {@link StatsLogger} implementation.
@@ -36,7 +37,7 @@ public class PrometheusStatsLogger implements StatsLogger {
     private final String scope;
     private final Map<String, String> labels;
 
-    PrometheusStatsLogger(PrometheusMetricsProvider provider, String scope, Map<String, String> labels) {
+    public PrometheusStatsLogger(PrometheusMetricsProvider provider, String scope, Map<String, String> labels) {
         this.provider = provider;
         this.scope = scope;
         this.labels = labels;
