+++ jetty-upgrade/bookkeeper-prometheus-metrics-provider/src/main/java/org/apache/pulsar/metrics/prometheus/bookkeeper/ThreadScopedDataSketchesStatsLogger.java
@@ -24,6 +24,7 @@ import java.util.concurrent.TimeUnit;
 import org.apache.bookkeeper.stats.OpStatsData;
 import org.apache.bookkeeper.stats.OpStatsLogger;
 import org.apache.bookkeeper.stats.ThreadRegistry;
+import org.apache.pulsar.bookkeeper.stats.prometheus.PrometheusMetricsProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
