+++ jetty-upgrade/bookkeeper-prometheus-metrics-provider/src/test/java/org/apache/pulsar/metrics/prometheus/bookkeeper/TestPrometheusMetricsProvider.java
@@ -33,6 +33,7 @@ import lombok.Cleanup;
 import org.apache.bookkeeper.stats.Counter;
 import org.apache.bookkeeper.stats.StatsLogger;
 import org.apache.commons.configuration2.PropertiesConfiguration;
+import org.apache.pulsar.bookkeeper.stats.prometheus.PrometheusMetricsProvider;
 import org.testng.annotations.Test;
 
 /**
